VIMS-328: Minor refactorMake ISA.PopulationSource a Long rather than Integer and remove BaseModel as StockRequirements' parent.

+++ modules/core/src/main/java/org/openlmis/core/domain/ProgramProductISA.java
@@ -94,11 +94,11 @@ public class ProgramProductISA extends BaseModel {
     isa.setAdjustmentValue(adjustmentValue);
   }
 
-  public Integer getPopulationSource() {
+  public Long getPopulationSource() {
     return isa.getPopulationSource();
   }
 
-  public void setPopulationSource(Integer sourceId) {
+  public void setPopulationSource(Long sourceId) {
     isa.setPopulationSource(sourceId);
   }
 }
